Refactored code base and transitioned to H2 database
The commit implements a major restructure and refactoring of the project. The Company, Contact, and related files were removed, and new classes for the "fridge" concept were added which includes food listings, diet restrictions, and claims. It also introduced AppUser to replace the Contact entity. The change replaces an old data schema related to Companies and Contacts to a new one more aligned with the fridge-sharing domain. Backend repositories and domain objects were updated accordingly. The database was also transitioned from an unspecified type to an H2 database. Additionally, frontend views related to creating and viewing listings were updated reflecting the new data schema. Tailwind CSS was introduced to the project's frontend for more streamlined styling. A postCSS config file was added for processing Tailwind CSS.