Update log warning level and improve food listing interface
This commit changes the logging level for the image compression message from info to warning. It also removes the null check for request expiry date and sets it to start of day by default. In the frontend, various enhancements have been made to the FoodListingDetailView including better placement of listings, claimant info and user control options. The default ClaimState has been changed from WAITING to UNCLAIMED, and a minor adjustment in the default data.sql file has been made to remove allergen 'NONE'.